fixed problem reading .wav files, remaining negative numbers

+++ audio-encoder/BitStream.cpp
@@ -18,13 +18,13 @@ BitStream::BitStream(string readname, string writename) {
 		remainingByteSlots = 0;
 }
 
-int BitStream::readBit () {
+short BitStream::readBit () {
 
 		if(readPosition >= writePosition) return -1;
 
 		ifstream* stream = new ifstream(readFilename->c_str(), ifstream::binary);
 
-		int bit = 0;
+		short bit = 0;
 		if (stream->is_open()) {
 
 				// get length of file
@@ -39,7 +39,7 @@ int BitStream::readBit () {
 
 				// bitwise
 				bit = (byteBuffer >> (7-readPosition%8)) & 0x1;
-				//int bit = (bytesBuffer[readPosition%8] >> (7-(readPosition%8))) & 0x1;
+				//short bit = (bytesBuffer[readPosition%8] >> (7-(readPosition%8))) & 0x1;
 				readPosition++;
 
 
@@ -48,15 +48,15 @@ int BitStream::readBit () {
 		return bit;
 }
 
-int* BitStream::readNBits(int nBits) {
+short* BitStream::readNBits(short nBits) {
 
-	int div = nBits/8;
-	int rem = nBits%8;
-	int bytesToRead = !rem ? div : div+1;
+	short div = nBits/8;
+	short rem = nBits%8;
+	short bytesToRead = !rem ? div : div+1;
 
 
 	ifstream* stream = new ifstream(readFilename->c_str());
-	int* bitBuffer = new int[nBits];
+	short* bitBuffer = new short[nBits];
 
 	if (stream->is_open()) {
 		// get length of file:
@@ -64,7 +64,7 @@ int* BitStream::readNBits(int nBits) {
 
 		char* bytesBuffer;		// read data as a block
 
-		int readUntilNow = 0;
+		short readUntilNow = 0;
 
 		if(readPosition%8 != 0) {
 			bytesToRead++;
@@ -73,9 +73,9 @@ int* BitStream::readNBits(int nBits) {
 			stream->read (bytesBuffer,bytesToRead);
 			stream->close();
 
-			int pos = readPosition%8;
+			short pos = readPosition%8;
 			while(pos < 8) {
-				int bit = (bytesBuffer[0] >> (7-pos++)) & 0x1;
+				short bit = (bytesBuffer[0] >> (7-pos++)) & 0x1;
 
 				bitBuffer[readUntilNow] = bit;			
 
@@ -100,11 +100,11 @@ int* BitStream::readNBits(int nBits) {
 
 		// bytwise reading
 		// if we already started reading bits, then skip to next byte
-		for(int i = (readUntilNow != 0); i < bytesToRead; i++) {
+		for(short i = (readUntilNow != 0); i < bytesToRead; i++) {
 			// bitwise reading
 
-			for(int j = 0; j < 8; j++) {
-				int bit = (bytesBuffer[i] >> (7-j)) & 0x1;
+			for(short j = 0; j < 8; j++) {
+				short bit = (bytesBuffer[i] >> (7-j)) & 0x1;
 
 				bitBuffer[readUntilNow] = bit;
 
@@ -126,15 +126,15 @@ int* BitStream::readNBits(int nBits) {
 	return bitBuffer;
 }
 
-void BitStream::writeBit(int* sequence) {
+void BitStream::writeBit(short* sequence) {
 
 	ofstream* stream = new ofstream();
 
 	unsigned char buffer = 0;
-	int bufferPos = 0;
+	short bufferPos = 0;
 
 	if(remainingByteSlots) {
-		int pos = 8 - remainingByteSlots;
+		short pos = 8 - remainingByteSlots;
 		surplusByte = surplusByte | (sequence[0] << (7-pos));
 		remainingByteSlots--;
 
@@ -155,23 +155,23 @@ void BitStream::writeBit(int* sequence) {
 
 }
 
-void BitStream::writeNBits(int nBits, int* sequence, int finalWrite) {
+void BitStream::writeNBits(short nBits, short* sequence, short finalWrite) {
 
 
 	ofstream* stream = new ofstream();
 	const char* filename = writeFilename->c_str();
 
 	char buffer = 0;
-	int bufferPos = 0;
+	short bufferPos = 0;
 
 	stream->open(filename, ios::out | ios::binary | ios::app);
 
-	int j = 0;
+	short j = 0;
 
 	stream->seekp(writePosition/8, stream->beg);
 	if(surplus) {
 
-		int pos = 8 - remainingByteSlots;
+		short pos = 8 - remainingByteSlots;
 
 		while(pos < 8) {		
 
@@ -195,8 +195,8 @@ void BitStream::writeNBits(int nBits, int* sequence, int finalWrite) {
 	/*if(j > 0) stream->seekp((writePosition/8)+1, stream->beg);
 	else stream->seekp((writePosition/8), stream->beg);*/
 
-	for(int i = j; i < nBits; i++) {
-		int bit = sequence[i];
+	for(short i = j; i < nBits; i++) {
+		short bit = sequence[i];
 
 		buffer = buffer | (bit << (7-bufferPos));
 	
@@ -235,7 +235,7 @@ void BitStream::writeNBits(int nBits, int* sequence, int finalWrite) {
 
 void BitStream::flush() {
 	cout << "FINAL: ";
-	for(int k = 0; k < 8; k++)
+	for(short k = 0; k < 8; k++)
 		cout << ((surplusByte >> (7-k)) & 0x1);
 	cout << "\n";
