+++ audio-encoder/BitStream.h
@@ -7,20 +7,20 @@ class BitStream {
 
 	public:
 		BitStream(string readname, string writename);
-		int readBit();	
-		int* readNBits(int nBits);
-		void writeBit(int* sequence);
-		void writeNBits(int nBits, int* sequence, int finalWrite);
+		short readBit();	
+		short* readNBits(short nBits);
+		void writeBit(short* sequence);
+		void writeNBits(short nBits, short* sequence, short finalWrite);
 
 	private:
 		string* readFilename;
 		string* writeFilename;
-		int readPosition;
-		int writePosition;
+		short readPosition;
+		short writePosition;
 
-		int surplus;
+		short surplus;
 		char surplusByte;
-		unsigned int remainingByteSlots;
+		unsigned short remainingByteSlots;
 
 		void flush();
 };
